requirements.txt
streamlit>=1.32.0
groq>=0.9.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
textstat>=0.7.3
wordcloud>=1.9.3
matplotlib>=3.8.0
nltk>=3.8.1
st-annotated-text>=4.0.1